move tabology to its own file
